+++ packages/html/src/index.ts
@@ -11,3 +11,4 @@ export { Redirect, redirect, redirectTo, redirectToNavRef } from './Redirect';
 export { NavRefRedirect } from './NavRefRedirect';
 export { NavLink, NavLinkProps, NavLinkParams } from './NavLink';
 export { BoundNavLink } from './BoundNavLink';
+export { CountingClock, CountingClockProps } from './CountingClock';