+++ ChangeLog.md
@@ -6,6 +6,7 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 
 ### Added
  - Support indicating the pending state on ToolTipAnt
+ - Add CountingClock widget to @moxb/html
 
 # [v0.2.0-beta.51](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.51) (2020-10-20)